+++ build_logic/src/main/kotlin/lambdynamiclights/task/AssembleNeoForgeJarTask.kt
@@ -4,7 +4,10 @@ import com.google.gson.JsonParser
 import dev.lambdaurora.mcdev.api.AccessWidenerToTransformer
 import dev.lambdaurora.mcdev.util.JsonUtils
 import dev.lambdaurora.mcdev.util.ZipFix
+import lambdynamiclights.Constants
 import org.gradle.api.file.RegularFileProperty
+import org.gradle.api.provider.Property
+import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputFile
 import org.gradle.api.tasks.TaskAction
 import java.nio.file.FileSystem
@@ -13,6 +16,9 @@ import java.nio.file.Files
 import javax.inject.Inject
 
 abstract class AssembleNeoForgeJarTask @Inject constructor() : AbstractAssembleJarTask() {
+	@get:Input
+	abstract val artifactGroup: Property<String>
+
 	@get:InputFile
 	abstract val runtimeMojmapJar: RegularFileProperty
 
@@ -21,6 +27,10 @@ abstract class AssembleNeoForgeJarTask @Inject constructor() : AbstractAssembleJ
 
 	@get:InputFile
 	abstract val jarJarMetadata: RegularFileProperty
+
+	init {
+		this.artifactGroup.convention(this.project.group as String)
+	}
 
 	@TaskAction
 	fun assemble() {
@@ -68,9 +78,7 @@ abstract class AssembleNeoForgeJarTask @Inject constructor() : AbstractAssembleJ
 				}
 			}
 
-			val jarjarDirPath = outFs.getPath("META-INF/jarjar")
-			this.createDirectories(jarjarDirPath)
-			this.copy(this.jarJarMetadata.get().asFile.toPath(), jarjarDirPath.resolve("metadata.json"))
+			this.handleJarJarMetadata(outFs)
 		}
 
 		ZipFix.makeZipReproducible(outputJar)
@@ -81,5 +89,23 @@ abstract class AssembleNeoForgeJarTask @Inject constructor() : AbstractAssembleJ
 			.asJsonObject
 		mixinsJson.addProperty("refmap", "lambdynlights-refmap.json")
 		this.writeString(fs.getPath("lambdynlights.mixins.json"), JsonUtils.GSON.toJson(mixinsJson))
+	}
+
+	private fun handleJarJarMetadata(fs: FileSystem) {
+		val jarjarDirPath = fs.getPath("META-INF/jarjar")
+		this.createDirectories(jarjarDirPath)
+
+		val metadata = JsonParser.parseString(Files.readString(this.jarJarMetadata.get().asFile.toPath())).asJsonObject
+		val jars = metadata.getAsJsonArray("jars")
+
+		for (jar in jars) {
+			val obj = jar.asJsonObject
+			val identifier = obj.getAsJsonObject("identifier")
+			if (identifier.getAsJsonPrimitive("group").asString.equals(this.artifactGroup.get())) {
+				identifier.addProperty("artifact", Constants.API_ARTIFACT)
+			}
+		}
+
+		Files.writeString(jarjarDirPath.resolve("metadata.json"), JsonUtils.GSON.toJson(metadata))
 	}
 }