+++ CHANGELOG.md
@@ -533,6 +533,12 @@
 - Updated Upside-down English translations.
 - Updated Turkish translations ([#326](https://github.com/LambdAurora/LambDynamicLights/pull/326)).
 
+### 4.8.4
+
+- Fixed bad API identifier in NeoForge JarJar metadata.
+  - This fixes some issues loading newer versions of LambDynamicLights with mods that include its API JAR on NeoForge.
+- Updated Vietnamese translations ([#330](https://github.com/LambdAurora/LambDynamicLights/pull/326)).
+
 [SpruceUI]: https://github.com/LambdAurora/SpruceUI "SpruceUI page"
 [pridelib]: https://github.com/Queerbric/pridelib "Pridelib page"
 [Yumi Commons]: https://github.com/YumiProject/yumi-commons "Yumi Commons page"